Creacion de CRUD y contexto parqa VolumenesDeProduccion

+++ PMMX.Modelo/Vistas/CRRView.cs
@@ -8,7 +8,7 @@ namespace PMMX.Modelo.Vistas
 {
     public class CRRView
     {
-        public int CRR_Total { get; set; }
+        public Double CRR_Total { get; set; }
         public int Objetivo { get; set; }
         public DateTime Fecha { get; set; }
     }
